requirements.txt
requests
pyyaml
python-dateutil
pymongo
python-gnupg